# address-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Address API documentation",
        "description": "API to manage addresses",
        "contact": {
            "email": "admin@curion.ch"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/addresses": {
            "get": {
                "tags": [
                    "Address"
                ],
                "summary": "Get list of addresses",
                "description": "Returns list of addresses",
                "operationId": "index",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AddressResourceCollection"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/addresses/{id}": {
            "get": {
                "tags": [
                    "Address"
                ],
                "summary": "Get address details",
                "description": "Returns an address by id",
                "operationId": "show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Address id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AddressResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/salutations": {
            "get": {
                "tags": [
                    "Salutation"
                ],
                "summary": "Get list of salutations",
                "description": "Returns list of salutations",
                "operationId": "index",
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SalutationResourceCollection"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "AddressResource": {
                "title": "AddressResource",
                "description": "Address resource",
                "required": [
                    "id",
                    "lastName",
                    "salutation"
                ],
                "properties": {
                    "id": {
                        "schema": "AddressResource",
                        "title": "id",
                        "description": "Id of the address",
                        "format": "int64",
                        "example": "1"
                    },
                    "createdAt": {
                        "schema": "AddressResource",
                        "title": "createdAt",
                        "description": "When the addres was created",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-06-06T16:28:28.070000Z"
                    },
                    "updatedAt": {
                        "schema": "AddressResource",
                        "title": "updatedAt",
                        "description": "When the addres was updated last",
                        "type": "string",
                        "format": "datetime",
                        "example": "2020-06-06T16:28:28.070000Z"
                    },
                    "firstName": {
                        "schema": "AddressResource",
                        "title": "firstName",
                        "description": "First name of the person",
                        "type": "string",
                        "example": "Max"
                    },
                    "lastName": {
                        "schema": "AddressResource",
                        "title": "lastName",
                        "description": "Last name of the person",
                        "type": "string",
                        "example": "Mustermann"
                    },
                    "address": {
                        "schema": "AddressResource",
                        "title": "address",
                        "description": "Street and house nr.",
                        "type": "string",
                        "example": "Barthelring 4c"
                    },
                    "postalCode": {
                        "schema": "AddressResource",
                        "title": "postalCode",
                        "description": "Postal code",
                        "type": "string",
                        "example": "71338"
                    },
                    "city": {
                        "schema": "AddressResource",
                        "title": "city",
                        "description": "City",
                        "type": "string",
                        "example": "Leipzig"
                    },
                    "birthday": {
                        "schema": "AddressResource",
                        "title": "birthday",
                        "description": "Birthday of the person",
                        "type": "string",
                        "format": "datetime",
                        "example": "2017-08-13T00:00:00.000000Z"
                    },
                    "salutation": {
                        "schema": "AddressResource",
                        "title": "salutation",
                        "description": "Salutation of the person",
                        "type": "string",
                        "example": "mr"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "AddressResource"
                }
            },
            "AddressResourceCollection": {
                "title": "AddressResourceCollection",
                "description": "Address resource collection",
                "required": [
                    "data"
                ],
                "properties": {
                    "data": {
                        "schema": "AddressResourceCollection",
                        "title": "Data",
                        "description": "Address collection",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/AddressResource"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "AddressResourceCollection"
                }
            },
            "SalutationResource": {
                "title": "SalutationResource",
                "description": "Salutation resource",
                "required": [
                    "id",
                    "key"
                ],
                "properties": {
                    "id": {
                        "schema": "SalutationResource",
                        "title": "id",
                        "description": "Id of the salutation",
                        "format": "int64",
                        "example": "1"
                    },
                    "key": {
                        "schema": "SalutationResource",
                        "title": "key",
                        "description": "Salutation key",
                        "example": "mr"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "SalutationResource"
                }
            },
            "SalutationResourceCollection": {
                "title": "SalutationResourceCollection",
                "description": "Salutation resource collection",
                "required": [
                    "data"
                ],
                "properties": {
                    "data": {
                        "schema": "SalutationResourceCollection",
                        "title": "Data",
                        "description": "Salutation collection",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/SalutationResource"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "SalutationResourceCollection"
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "apiKey",
                "description": "Bearer JWT Auth",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "tags": [
        {
            "name": "Address",
            "description": "API Endpoints of Address"
        },
        {
            "name": "Salutation",
            "description": "API Endpoints of Salutation"
        }
    ]
}